faker_data/requirements.txt
Faker==13.3.0
psycopg2_binary==2.9.6
